:truck: move unit tests to test/unit
